feat: Automatically include the resource group for the requested object
When printing an object & if there exists another object in the node map
that has the same kind but belongs to a different group (eg. "services.v1"
vs "services.v1.serving.knative.dev"), we prepend the object's name with
its GroupKind instead of its Kind to clearly indicate which resource type
it belongs to.

Signed-off-by: Justin Toh <tohjustin@hotmail.com>